Contact role dict (#51)
This PR creates a new column `contact_role_dicts`. The structure within the column is the following:
```
{
  role_1: [
    "contact_1",
    "contact_2"
  ],
  role_1_keyids: [
    "nbcontact contact_1",
    "nbcontact contact_2"
  ]
}
```
The column can be combined with the Get a specific Array Element filter in icinga to get the list of contact with a specific role. Background is the separation between f.e. sms and mail notifications. Combined with these scripts the contacts can be taken out of a variable and be used for the notification.